Fix crash when reloading arrays
There was an error in `Array1D` where when the data were unmapped, it
"invalidated" the associated Viskores `UnknownArrayHandle` by releasing
the resources. However, this does not actually invalidate the array. It
just sets the size to zero. The next time the array was set and used,
it registered as a valid array and therefore skipped pulling the new
array.

This fix properly invalidates the `UnknownArrayHandle` by resetting it
to a new object. This will automatically release resources of the array
is not still used elsewhere.